Harmonize description representation
Description was previously stored "raw" in the XML code, resulting in
plenty of disruption with certain special characters (#46). On the other
hand it is read in PathwayData::getWikiDescription() as some escaped
text string, causing problems with round-tripability.

This commit modifies the writing part to escape the text string; a write
to the "nodeValue" property now writes to "textcontent" instead. Doing
so makes the description storage more robust and internally consistent
with the reading part. Writing to "textcontent" requires PHP 5.6.1+.

Fix #46.